requirements.txt
aes==1.2.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
async==0.6.2
attrs==23.1.0
bcrypt==4.1.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored==2.2.3
console==0.9907
crypto==1.4.1
cryptography==41.0.7
dnspython==2.4.2
ezenv==0.92
frozenlist==1.4.0
futures==3.0.5
geoip2==4.7.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
markdown-it-py==3.0.0
maxminddb==2.5.1
mdurl==0.1.2
multidict==6.0.4
Naked==0.1.32
paramiko==3.3.1
Pillow==10.1.0
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
scapy==2.5.0
serde==0.9.0
setuptools==69.0.2
shellescape==3.8.1
sniffio==1.3.0
so==0.5
soupsieve==2.5
tabulate==0.9.0
termcolor==2.3.0
urllib3==2.1.0
yarl==1.9.3
